backend/requirements.txt
huggingface_hub
torch
transformers
sentence-transformers
Pillow
numpy
uvicorn
fastapi
python-dotenv